Implemented role validation for deletion service; Rebased blacklist functionality to the JwtService; Critical bug fixes and improvements.

+++ src/main/java/com/portfolio/blog/services/impl/PostService.java
@@ -3,6 +3,8 @@ package com.portfolio.blog.services.impl;
 import com.portfolio.blog.domain.dto.post.PostRequest;
 import com.portfolio.blog.domain.dto.post.PostResponse;
 import com.portfolio.blog.domain.entities.PostEntity;
+import com.portfolio.blog.domain.entities.Role;
+import com.portfolio.blog.domain.entities.UserEntity;
 import com.portfolio.blog.mappers.PostMapper;
 import com.portfolio.blog.repositories.PostRepository;
 import com.portfolio.blog.services.CategoryServiceInterface;
@@ -12,12 +14,15 @@ import com.portfolio.blog.services.UserServiceInterface;
 import jakarta.persistence.EntityNotFoundException;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
 
+@Slf4j
 @Service
 @RequiredArgsConstructor
 public class PostService implements PostServiceInterface {
@@ -58,7 +63,9 @@ public class PostService implements PostServiceInterface {
         PostEntity postEntity = mapper.requestToEntity(request);
 
         postEntity.setAuthor(userService.extractUserFromSecurityContextHolder());
-        postEntity.setReadingTime(postEntity.getContent().length() / 2);
+        postEntity.setReadingTime(
+                calculateReadingTime(request.getContent()
+                ));
         postEntity.setCategory(categoryService.verifyCategory(request));
         postEntity.setTags(tagService.verifyTags(request));
 
@@ -91,7 +98,29 @@ public class PostService implements PostServiceInterface {
 
     @Override
     public void delete(UUID uuid) {
-        repository.deleteById(uuid);
+
+        //  If Role.ADMIN user is allowed to delete posts of any user.
+        //  In case if Role.USER, user is only allowed to delete his own posts.
+
+        Optional<PostEntity> postToDelete = repository.findById(uuid);
+        if(postToDelete.isEmpty()) throw new EntityNotFoundException();
+        UUID authorUUID = postToDelete.get().getAuthor().getId();
+
+        UserEntity currentUser = userService.extractUserFromSecurityContextHolder();
+
+        boolean canDelete = currentUser.getRole().equals(Role.ADMIN)
+                || currentUser.getId().equals(authorUUID) && currentUser.getRole().equals(Role.USER);
+
+        if(canDelete) {
+
+            repository.deleteById(uuid);
+            log.info(currentUser.getUsername() + " has deleted post: " + postToDelete.get().getTitle());
+
+        } else {
+            log.warn(currentUser.getEmail() + " attempted to delete post of other user ");
+            throw new AccessDeniedException("");
+        }
+
     }
 
     private int calculateReadingTime(String content) {
